refactor images in case list to use new inflation method; create distinction between image util methods that are for images being saved (in FileUtils) and images being displayed (MediaUtil)
